+++ packages/core/src/components/AvatarGroup/AvatarGroup.tsx
@@ -4,19 +4,10 @@ import { AvatarProps } from "../Avatar/Avatar";
 import AvatarGroupCounter from "./AvatarGroupCounter";
 import VibeComponentProps from "../../types/VibeComponentProps";
 import { AvatarSize, AvatarType } from "../Avatar/Avatar.types";
-import { CounterColor } from "../Counter/Counter.types";
 import { avatarOnClick } from "./AvatarGroupHelper";
 import { TooltipProps } from "../Tooltip/Tooltip";
 import styles from "./AvatarGroup.module.scss";
-
-export type AvatarGroupCounterVisualProps = {
-  color?: Extract<CounterColor, "light" | "dark">;
-  count?: number;
-  prefix?: string;
-  maxDigits?: number;
-  ariaLabelItemsName?: string;
-  noAnimation?: boolean;
-};
+import { AvatarGroupCounterVisualProps } from "./AvatarGroup.types";
 
 export interface AvatarGroupProps extends VibeComponentProps {
   avatarClassName?: string;
