+++ packages/core/src/components/AvatarGroup/AvatarGroup.types.ts
@@ -1,0 +1,10 @@
+import { CounterColor } from "../Counter/Counter.types";
+
+export type AvatarGroupCounterVisualProps = {
+  color?: Extract<CounterColor, "light" | "dark">;
+  count?: number;
+  prefix?: string;
+  maxDigits?: number;
+  ariaLabelItemsName?: string;
+  noAnimation?: boolean;
+};
