+++ packages/core/src/components/AvatarGroup/index.ts
@@ -1,1 +1,2 @@
 export { default as AvatarGroup, AvatarGroupProps } from "./AvatarGroup";
+export * from "./AvatarGroup.types";
