+++ packages/core/src/components/Divider/Divider.types.ts
@@ -1,1 +1,1 @@
-export type DirectionType = "vertical" | "horizontal";
+export type DividerDirection = "vertical" | "horizontal";
