+++ packages/core/src/components/Link/Link.types.ts
@@ -1,3 +1,3 @@
 export type LinkTarget = "_blank" | "_self" | "_parent" | "_top";
 
-export type IconPosition = "start" | "end";
+export type LinkIconPosition = "start" | "end";
