+++ packages/core/src/components/VirtualizedGrid/VirtualizedGrid.types.ts
@@ -1,0 +1,6 @@
+export type VirtualizedGridItemType = {
+  value: string;
+  height: number;
+  width: number;
+  id: string;
+};
