+++ packages/core/src/components/VirtualizedGrid/index.ts
@@ -1,1 +1,2 @@
 export { default as VirtualizedGrid, VirtualizedGridProps } from "./VirtualizedGrid";
+export * from "./VirtualizedGrid.types";
