+++ packages/core/src/components/VirtualizedList/VirtualizedList.tsx
@@ -13,8 +13,6 @@ import React, {
 } from "react";
 import { noop as NOOP } from "lodash-es";
 import {
-  Layout,
-  ScrollDirection,
   VariableSizeList as List,
   ListOnItemsRenderedProps,
   ListChildComponentProps,
@@ -36,14 +34,7 @@ import { ComponentDefaultTestId } from "../../tests/constants";
 import VibeComponentProps from "src/types/VibeComponentProps";
 import VibeComponent from "../../types/VibeComponent";
 import styles from "./VirtualizedList.module.scss";
-
-export type VirtualizedListItem = {
-  value?: string | Record<string, unknown>;
-  height?: number;
-  width?: number;
-  id?: string;
-  offsetTop?: number;
-};
+import { VirtualizedListItem, VirtualizedListLayout, VirtualizedListScrollDirection } from "./VirtualizedList.types";
 
 export interface VirtualizedListProps extends VibeComponentProps {
   /**
@@ -56,7 +47,7 @@ export interface VirtualizedListProps extends VibeComponentProps {
    * - "vertical" (default) - Up/down scrolling.
    * - "horizontal" - Left/right scrolling.
    */
-  layout?: Layout;
+  layout?: VirtualizedListLayout;
   /**
    * A list of items to be rendered
    */
@@ -136,7 +127,11 @@ export interface VirtualizedListProps extends VibeComponentProps {
    */
   scrollToId?: string;
   virtualListRef?: ForwardedRef<HTMLElement>;
-  onScroll?: (horizontalScrollDirection: ScrollDirection, scrollTop: number, scrollUpdateWasRequested: boolean) => void;
+  onScroll?: (
+    horizontalScrollDirection: VirtualizedListScrollDirection,
+    scrollTop: number,
+    scrollUpdateWasRequested: boolean
+  ) => void;
 }
 
 const VirtualizedList: VibeComponent<VirtualizedListProps> = forwardRef(
