+++ packages/core/src/components/VirtualizedList/VirtualizedList.types.ts
@@ -1,0 +1,9 @@
+export type VirtualizedListItem = {
+  value?: string | Record<string, unknown>;
+  height?: number;
+  width?: number;
+  id?: string;
+  offsetTop?: number;
+};
+
+export { Layout as VirtualizedListLayout, ScrollDirection as VirtualizedListScrollDirection } from "react-window";
