+++ packages/core/src/components/VirtualizedList/index.ts
@@ -1,1 +1,3 @@
 export { default as VirtualizedList, VirtualizedListProps } from "./VirtualizedList";
+
+export * from "./VirtualizedList.types";
